+++ src/test/resources/apis/v3/callout-http-invalid-target.json
