+++ src/test/resources/apis/v3/callout-http-post-with-accents.json
